feat: new file added for shared components
